Use the "on develop" phase not cpanfile.dzil
Because although the other doesn't install stuff by default, we can
set it not to install those things in the Makefile.

I do know that the other way we could have added Makefile targets to
handle automatically enabling the develop/dzil deps when doing
dist-zilla things, but this seems cleaner.