Remove the Git::Contributors Plugin
Concern was expressed at putting these names and email addresses in
places that will be indexed easily by web scrapers.  This will remove it
from the README.md that shows up on GitHub, the POD for the module, and
the meta.yml/meta.json that the MetaCPAN uses to show those names in the
sidebar.