Try to find the correct remote for github
Some of the Dist::Zilla features we are using require a remote exists
for GitHub, this looks for a remote in `git remote -v` that looks like
github.com and uses it.  They all default to "origin", but that often
points to our internal github mirror when making changes, this should
allow it to have a different name.